clean up

+++ FRED/trainers.py
@@ -59,7 +59,7 @@ class Trainer(object):
                 # have model compute losses, compile them into cost using loss weights
                 if self.data_type == "Flow Neighbor":
                     data['X'] = data['X'].float().to(self.device) # We convert to float32s for compatibility with apple mps
-                    data['neighbors'] = data['neighbors'].float().to(self.device)
+                    data['neighbors'] = data['neighbors'].to(self.device)
                     # data['P'] = data['P'].to(self.device)
                     data['precomputed distances'] = data['precomputed distances'].float().to(self.device)
                 elif self.data_type == "Flow Prediction":
